+++ R/compile_pseudo_pop.R
@@ -63,6 +63,7 @@
 #'                          matching_fun = "matching_l1",
 #'                          covar_bl_method = 'absolute',
 #'                          covar_bl_trs = 0.1,
+#'                          covar_bl_trs_type= "mean",
 #'                          delta_n = 0.5,
 #'                          scale = 1)
 #'